ビルド用の設定を追加

+++ scripts/build-chat-log.py
@@ -33,8 +33,8 @@ PATTERN_INPUT_CSS = [
 def initFile(path):
     if os.path.isfile(path):
         os.remove(path)
-    with open(path, "w", encoding="UTF-8") as f:
-        pass
+        with open(path, "w", encoding="UTF-8") as f:
+            pass
 
 # --------------------------------------------------
 # 正規表現に引っかかったら、そこに書いてあるパスのファイルの内容を整形して返却する
@@ -47,44 +47,48 @@ def getIncludeText(regInfoArr, line):
         if (result):
             # 対象ファイルのパスを取得
             path = "dist/" + result.group(1)
-            with open(path) as fr:
-                # ファイルの内容を全取得
-                lines = fr.read()
-                # テンプレートに流し込んで返却
-                return regInfo[1].replace("\1", lines)
+            if os.path.exists(path):
+              with open(path) as fr:
+                  # ファイルの内容を全取得
+                  lines = fr.read()
+                  # テンプレートに流し込んで返却
+                  return regInfo[1].replace("\1", lines)
     # 1つも引っかからなければ元のテキストを返却
     return line
 
-# --------------------------------------------------
-# キレイに行に分解してあげる
-with open(BASE_HTML_ORG) as fr:
-    lines = fr.read()
-    lines = re.sub('(?!<br)(<[^/!])', r'\n\1', lines)
-    lines = re.sub('(<\/body>)', r'\n\1', lines)
-    lines = re.sub('(<\/head>)', r'\n\1', lines)
-    lines = re.sub('<link.+ rel=\"?preload\"?.*>\n', '', lines)
-    with open(BASE_HTML, mode='w') as fw:
-        fw.write(lines)
+if os.path.exists(BASE_HTML_ORG):
+  # --------------------------------------------------
+  # キレイに行に分解してあげる
+  with open(BASE_HTML_ORG) as fr:
+      lines = fr.read()
+      lines = re.sub('(?!<br)(<[^/!])', r'\n\1', lines)
+      lines = re.sub('(<\/body>)', r'\n\1', lines)
+      lines = re.sub('(<\/head>)', r'\n\1', lines)
+      lines = re.sub('<link.+ rel=\"?preload\"?.*>\n', '', lines)
+      with open(BASE_HTML, mode='w') as fw:
+          fw.write(lines)
 
 # --------------------------------------------------
 # 成果物を出力する
 initFile(OUTPUT_HTML)
-with open(BASE_HTML) as fr:
-    line = fr.readline()
-    while line:
-        # キレイな行の単位
-        # print(line, end="")
-        convertedStr = getIncludeText([PATTERN_INPUT_SCRIPT, PATTERN_INPUT_CSS], line)
-        convertedStr = re.sub(r'^\@charset \"UTF-8\";', '', convertedStr, flags=re.MULTILINE)
-        with open(OUTPUT_HTML, mode='a') as fa:
-            print(convertedStr, file=fa, end="")
-        line = fr.readline()
+if os.path.exists(BASE_HTML):
+  with open(BASE_HTML) as fr:
+      line = fr.readline()
+      while line:
+          # キレイな行の単位
+          # print(line, end="")
+          convertedStr = getIncludeText([PATTERN_INPUT_SCRIPT, PATTERN_INPUT_CSS], line)
+          convertedStr = re.sub(r'^\@charset \"UTF-8\";', '', convertedStr, flags=re.MULTILINE)
+          with open(OUTPUT_HTML, mode='a') as fa:
+              print(convertedStr, file=fa, end="")
+          line = fr.readline()
 
-# --------------------------------------------------
-# 後処理
+  # --------------------------------------------------
+  # 後処理
 
-# 中間ファイルを削除する
-os.remove(BASE_HTML)
+  # 中間ファイルを削除する
+  os.remove(BASE_HTML)
 
 # 逆輸入
-shutil.copyfile(OUTPUT_HTML, DEVELOP_HTML)
+if os.path.exists(OUTPUT_HTML):
+    shutil.copyfile(OUTPUT_HTML, DEVELOP_HTML)
